pyproject.toml
[tool.poetry]
name = "agent-with-memory"
version = "0.1.0"
description = "testing an agent with memory"
authors = ["hainghiem <hainghiem375@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "agent_with_memory"}]

[tool.poetry.dependencies]
python = "^3.10"
autopep8 = "*"
langchain = "*"
python-dotenv = "*"
gunicorn = "*"
uvicorn = "*"
openai = "*"
faiss-cpu = "*"
fastapi = "*"
unstructured = "*"
tiktoken = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
